requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
PyPDF2==3.0.1
reportlab==4.0.4
Pillow==10.0.1
gunicorn==21.2.0 